Allow overwriting the final report generation
