+++ src/lib.rs
@@ -10,8 +10,7 @@
 
 use bstr::ByteSlice;
 pub use color_eyre;
-use color_eyre::eyre::{bail, Context, Result};
-use colored::*;
+use color_eyre::eyre::{Context, Result};
 use crossbeam_channel::unbounded;
 use parser::{ErrorMatch, Pattern, Revisioned};
 use regex::bytes::Regex;
@@ -21,7 +20,6 @@ use std::borrow::Cow;
 use std::collections::{HashSet, VecDeque};
 use std::ffi::OsString;
 use std::fmt::Display;
-use std::fmt::Write;
 use std::num::NonZeroUsize;
 use std::path::{Path, PathBuf};
 use std::process::{Command, ExitStatus};
@@ -539,194 +537,7 @@ pub fn run_tests_generic(
         }
     }
 
-    // Print all errors in a single thread to show reliable output
-    if !failures.is_empty() {
-        for (path, cmd, revision, errors, stderr) in &failures {
-            let _guard = status_emitter.failed_test(revision, path, cmd, stderr);
-            for error in errors {
-                print_error(error, &path.display().to_string(), &revision);
-            }
-        }
-        eprintln!("{}", "FAILURES:".red().underline().bold());
-        for (path, _cmd, _revision, _errors, _stderr) in &failures {
-            eprintln!("    {}", path.display());
-        }
-        eprintln!();
-        eprintln!(
-            "test result: {}. {} tests failed, {} tests passed, {} ignored, {} filtered out",
-            "FAIL".red(),
-            failures.len().to_string().red().bold(),
-            succeeded.to_string().green(),
-            ignored.to_string().yellow(),
-            filtered.to_string().yellow(),
-        );
-        bail!("tests failed");
-    }
-    eprintln!();
-    eprintln!(
-        "test result: {}. {} tests passed, {} ignored, {} filtered out",
-        "ok".green(),
-        succeeded.to_string().green(),
-        ignored.to_string().yellow(),
-        filtered.to_string().yellow(),
-    );
-    eprintln!();
-    Ok(())
-}
-
-fn print_error(error: &Error, path: &str, revision: &str) {
-    match error {
-        Error::ExitStatus {
-            mode,
-            status,
-            expected,
-        } => {
-            github_actions::error(
-                path,
-                format!("{mode} test{revision} got {status}, but expected {expected}"),
-            );
-            eprintln!("{mode} test got {status}, but expected {expected}")
-        }
-        Error::Command { kind, status } => {
-            github_actions::error(path, format!("{kind}{revision} failed with {status}"));
-            eprintln!("{kind} failed with {status}");
-        }
-        Error::PatternNotFound {
-            pattern,
-            definition_line,
-        } => {
-            github_actions::error(path, format!("Pattern not found{revision}"))
-                .line(*definition_line);
-            match pattern {
-                Pattern::SubString(s) => {
-                    eprintln!("substring `{s}` {} in stderr output", "not found".red())
-                }
-                Pattern::Regex(r) => {
-                    eprintln!("`/{r}/` does {} stderr output", "not match".red())
-                }
-            }
-            eprintln!(
-                "expected because of pattern here: {}",
-                format!("{path}:{definition_line}").bold()
-            );
-        }
-        Error::NoPatternsFound => {
-            github_actions::error(
-                path,
-                format!("no error patterns found in fail test{revision}"),
-            );
-            eprintln!("{}", "no error patterns found in fail test".red());
-        }
-        Error::PatternFoundInPassTest => {
-            github_actions::error(path, format!("error pattern found in pass test{revision}"));
-            eprintln!("{}", "error pattern found in pass test".red())
-        }
-        Error::OutputDiffers {
-            path: output_path,
-            actual,
-            expected,
-        } => {
-            let mut err = github_actions::error(
-                if expected.is_empty() {
-                    path.to_owned()
-                } else {
-                    output_path.display().to_string()
-                },
-                "actual output differs from expected",
-            );
-            writeln!(err, "```diff").unwrap();
-            let mut seen_diff_line = Some(0);
-            for r in ::diff::lines(expected.to_str().unwrap(), actual.to_str().unwrap()) {
-                if let Some(line) = &mut seen_diff_line {
-                    *line += 1;
-                }
-                let mut seen_diff = || {
-                    if let Some(line) = seen_diff_line.take() {
-                        writeln!(err, "{line} unchanged lines skipped").unwrap();
-                    }
-                };
-                match r {
-                    ::diff::Result::Both(l, r) => {
-                        if l != r {
-                            seen_diff();
-                            writeln!(err, "-{l}").unwrap();
-                            writeln!(err, "+{r}").unwrap();
-                        } else if seen_diff_line.is_none() {
-                            writeln!(err, " {l}").unwrap()
-                        }
-                    }
-                    ::diff::Result::Left(l) => {
-                        seen_diff();
-                        writeln!(err, "-{l}").unwrap();
-                    }
-                    ::diff::Result::Right(r) => {
-                        seen_diff();
-                        writeln!(err, "+{r}").unwrap();
-                    }
-                }
-            }
-            writeln!(err, "```").unwrap();
-            eprintln!("{}", "actual output differed from expected".underline());
-            eprintln!("{}", format!("--- {}", output_path.display()).red());
-            eprintln!("{}", "+++ <stderr output>".green());
-            diff::print_diff(expected, actual);
-        }
-        Error::ErrorsWithoutPattern { path: None, msgs } => {
-            eprintln!(
-        "There were {} unmatched diagnostics that occurred outside the testfile and had no pattern",
-        msgs.len(),
-    );
-            for Message { level, message } in msgs {
-                eprintln!("    {level:?}: {message}")
-            }
-            let mut err = github_actions::error(
-                path,
-                format!("Unmatched diagnostics outside the testfile{revision}"),
-            );
-            for Message { level, message } in msgs {
-                writeln!(err, "{level:?}: {message}").unwrap();
-            }
-        }
-        Error::ErrorsWithoutPattern {
-            path: Some((path, line)),
-            msgs,
-        } => {
-            let path = path.display();
-            eprintln!(
-                "There were {} unmatched diagnostics at {path}:{line}",
-                msgs.len(),
-            );
-            for Message { level, message } in msgs {
-                eprintln!("    {level:?}: {message}")
-            }
-            let mut err = github_actions::error(&path, format!("Unmatched diagnostics{revision}"))
-                .line(*line);
-            for Message { level, message } in msgs {
-                writeln!(err, "{level:?}: {message}").unwrap();
-            }
-        }
-        Error::InvalidComment { msg, line } => {
-            let mut err =
-                github_actions::error(path, format!("Could not parse comment")).line(*line);
-            writeln!(err, "{msg}").unwrap();
-            eprintln!("Could not parse comment in {path}:{line} because\n{msg}",)
-        }
-        Error::Bug(msg) => {
-            eprintln!("A bug in `ui_test` occurred: {msg}");
-        }
-        Error::Aux {
-            path: aux_path,
-            errors,
-            line,
-        } => {
-            eprintln!("Aux build from {path}:{line} failed");
-            github_actions::error(path, format!("Aux build failed")).line(*line);
-            for error in errors {
-                print_error(error, &aux_path.display().to_string(), "");
-            }
-        }
-    }
-    eprintln!();
+    status_emitter.finalize(&failures, succeeded, ignored, filtered)
 }
 
 fn parse_and_test_file(path: &Path, config: &Config) -> Vec<TestRun> {
