+++ src/status_emitter.rs
@@ -1,9 +1,18 @@
 //! Variaous schemes for reporting messages during testing or after testing is done.
 
+use bstr::ByteSlice;
+use color_eyre::eyre::bail;
 use colored::Colorize;
 
-use crate::{github_actions, Config, TestResult};
-use std::{fmt::Debug, io::Write, path::Path, process::Command};
+use crate::{
+    github_actions, parser::Pattern, rustc_stderr::Message, Config, Error, Result, TestResult,
+};
+use std::{
+    fmt::{Debug, Write as _},
+    io::Write as _,
+    path::{Path, PathBuf},
+    process::Command,
+};
 
 /// A generic way to handle the output of this crate.
 pub trait StatusEmitter: Sync {
@@ -19,12 +28,22 @@ pub trait StatusEmitter: Sync {
 
     /// Start a test run and return a handle for reporting individual tests' results
     fn run_tests(&self, config: &Config) -> Box<dyn DuringTestRun>;
+    /// Create a report about the entire test run at the end.
+    fn finalize(
+        &self,
+        _failures: &[(PathBuf, Command, String, Vec<Error>, Vec<u8>)],
+        _succeeded: usize,
+        _ignored: usize,
+        _filtered: usize,
+    ) -> Result<()> {
+        Ok(())
+    }
 }
 
 /// Report information during test runs.
 pub trait DuringTestRun {
     /// A test has finished, handle the result.
-    fn test_result(&mut self, path: &Path, revision: &str, result: &TestResult);
+    fn test_result(&mut self, _path: &Path, _revision: &str, _result: &TestResult) {}
 }
 
 /// A human readable output emitter.
@@ -71,6 +90,255 @@ impl StatusEmitter for Text {
             Box::new(Text)
         }
     }
+
+    fn finalize(
+        &self,
+        failures: &[(PathBuf, Command, String, Vec<Error>, Vec<u8>)],
+        succeeded: usize,
+        ignored: usize,
+        filtered: usize,
+    ) -> Result<()> {
+        // Print all errors in a single thread to show reliable output
+        if !failures.is_empty() {
+            for (path, cmd, revision, errors, stderr) in failures {
+                let _guard = self.failed_test(revision, path, cmd, stderr);
+                for error in errors {
+                    print_error(error, &path.display().to_string());
+                }
+            }
+            eprintln!("{}", "FAILURES:".red().underline().bold());
+            for (path, _cmd, _revision, _errors, _stderr) in failures {
+                eprintln!("    {}", path.display());
+            }
+            eprintln!();
+            eprintln!(
+                "test result: {}. {} tests failed, {} tests passed, {} ignored, {} filtered out",
+                "FAIL".red(),
+                failures.len().to_string().red().bold(),
+                succeeded.to_string().green(),
+                ignored.to_string().yellow(),
+                filtered.to_string().yellow(),
+            );
+            bail!("tests failed");
+        }
+        eprintln!();
+        eprintln!(
+            "test result: {}. {} tests passed, {} ignored, {} filtered out",
+            "ok".green(),
+            succeeded.to_string().green(),
+            ignored.to_string().yellow(),
+            filtered.to_string().yellow(),
+        );
+        eprintln!();
+        Ok(())
+    }
+}
+
+fn print_error(error: &Error, path: &str) {
+    match error {
+        Error::ExitStatus {
+            mode,
+            status,
+            expected,
+        } => {
+            eprintln!("{mode} test got {status}, but expected {expected}")
+        }
+        Error::Command { kind, status } => {
+            eprintln!("{kind} failed with {status}");
+        }
+        Error::PatternNotFound {
+            pattern,
+            definition_line,
+        } => {
+            match pattern {
+                Pattern::SubString(s) => {
+                    eprintln!("substring `{s}` {} in stderr output", "not found".red())
+                }
+                Pattern::Regex(r) => {
+                    eprintln!("`/{r}/` does {} stderr output", "not match".red())
+                }
+            }
+            eprintln!(
+                "expected because of pattern here: {}",
+                format!("{path}:{definition_line}").bold()
+            );
+        }
+        Error::NoPatternsFound => {
+            eprintln!("{}", "no error patterns found in fail test".red());
+        }
+        Error::PatternFoundInPassTest => {
+            eprintln!("{}", "error pattern found in pass test".red())
+        }
+        Error::OutputDiffers {
+            path: output_path,
+            actual,
+            expected,
+        } => {
+            eprintln!("{}", "actual output differed from expected".underline());
+            eprintln!("{}", format!("--- {}", output_path.display()).red());
+            eprintln!("{}", "+++ <stderr output>".green());
+            crate::diff::print_diff(expected, actual);
+        }
+        Error::ErrorsWithoutPattern { path: None, msgs } => {
+            eprintln!(
+                "There were {} unmatched diagnostics that occurred outside the testfile and had no pattern",
+                msgs.len(),
+            );
+            for Message { level, message } in msgs {
+                eprintln!("    {level:?}: {message}")
+            }
+        }
+        Error::ErrorsWithoutPattern {
+            path: Some((path, line)),
+            msgs,
+        } => {
+            let path = path.display();
+            eprintln!(
+                "There were {} unmatched diagnostics at {path}:{line}",
+                msgs.len(),
+            );
+            for Message { level, message } in msgs {
+                eprintln!("    {level:?}: {message}")
+            }
+        }
+        Error::InvalidComment { msg, line } => {
+            let mut err =
+                github_actions::error(path, format!("Could not parse comment")).line(*line);
+            writeln!(err, "{msg}").unwrap();
+            eprintln!("Could not parse comment in {path}:{line} because\n{msg}",)
+        }
+        Error::Bug(msg) => {
+            eprintln!("A bug in `ui_test` occurred: {msg}");
+        }
+        Error::Aux {
+            path: aux_path,
+            errors,
+            line,
+        } => {
+            eprintln!("Aux build from {path}:{line} failed");
+            for error in errors {
+                print_error(error, &aux_path.display().to_string());
+            }
+        }
+    }
+    eprintln!();
+}
+
+fn gha_error(error: &Error, path: &str, revision: &str) {
+    match error {
+        Error::ExitStatus {
+            mode,
+            status,
+            expected,
+        } => {
+            github_actions::error(
+                path,
+                format!("{mode} test{revision} got {status}, but expected {expected}"),
+            );
+        }
+        Error::Command { kind, status } => {
+            github_actions::error(path, format!("{kind}{revision} failed with {status}"));
+        }
+        Error::PatternNotFound {
+            pattern: _,
+            definition_line,
+        } => {
+            github_actions::error(path, format!("Pattern not found{revision}"))
+                .line(*definition_line);
+        }
+        Error::NoPatternsFound => {
+            github_actions::error(
+                path,
+                format!("no error patterns found in fail test{revision}"),
+            );
+        }
+        Error::PatternFoundInPassTest => {
+            github_actions::error(path, format!("error pattern found in pass test{revision}"));
+        }
+        Error::OutputDiffers {
+            path: output_path,
+            actual,
+            expected,
+        } => {
+            let mut err = github_actions::error(
+                if expected.is_empty() {
+                    path.to_owned()
+                } else {
+                    output_path.display().to_string()
+                },
+                "actual output differs from expected",
+            );
+            writeln!(err, "```diff").unwrap();
+            let mut seen_diff_line = Some(0);
+            for r in ::diff::lines(expected.to_str().unwrap(), actual.to_str().unwrap()) {
+                if let Some(line) = &mut seen_diff_line {
+                    *line += 1;
+                }
+                let mut seen_diff = || {
+                    if let Some(line) = seen_diff_line.take() {
+                        writeln!(err, "{line} unchanged lines skipped").unwrap();
+                    }
+                };
+                match r {
+                    ::diff::Result::Both(l, r) => {
+                        if l != r {
+                            seen_diff();
+                            writeln!(err, "-{l}").unwrap();
+                            writeln!(err, "+{r}").unwrap();
+                        } else if seen_diff_line.is_none() {
+                            writeln!(err, " {l}").unwrap()
+                        }
+                    }
+                    ::diff::Result::Left(l) => {
+                        seen_diff();
+                        writeln!(err, "-{l}").unwrap();
+                    }
+                    ::diff::Result::Right(r) => {
+                        seen_diff();
+                        writeln!(err, "+{r}").unwrap();
+                    }
+                }
+            }
+            writeln!(err, "```").unwrap();
+        }
+        Error::ErrorsWithoutPattern { path: None, msgs } => {
+            let mut err = github_actions::error(
+                path,
+                format!("Unmatched diagnostics outside the testfile{revision}"),
+            );
+            for Message { level, message } in msgs {
+                writeln!(err, "{level:?}: {message}").unwrap();
+            }
+        }
+        Error::ErrorsWithoutPattern {
+            path: Some((path, line)),
+            msgs,
+        } => {
+            let path = path.display();
+            let mut err = github_actions::error(&path, format!("Unmatched diagnostics{revision}"))
+                .line(*line);
+            for Message { level, message } in msgs {
+                writeln!(err, "{level:?}: {message}").unwrap();
+            }
+        }
+        Error::InvalidComment { msg, line } => {
+            let mut err =
+                github_actions::error(path, format!("Could not parse comment")).line(*line);
+            writeln!(err, "{msg}").unwrap();
+        }
+        Error::Bug(_) => {}
+        Error::Aux {
+            path: aux_path,
+            errors,
+            line,
+        } => {
+            github_actions::error(path, format!("Aux build failed")).line(*line);
+            for error in errors {
+                gha_error(error, &aux_path.display().to_string(), "")
+            }
+        }
+    }
+    eprintln!();
 }
 
 impl DuringTestRun for Text {
@@ -132,6 +400,27 @@ impl StatusEmitter for Gha {
 
     fn run_tests(&self, _config: &Config) -> Box<dyn DuringTestRun> {
         Box::new(Gha)
+    }
+
+    fn finalize(
+        &self,
+        failures: &[(PathBuf, Command, String, Vec<Error>, Vec<u8>)],
+        _succeeded: usize,
+        _ignored: usize,
+        _filtered: usize,
+    ) -> Result<()> {
+        for (path, cmd, revision, errors, stderr) in failures {
+            let revision = if revision.is_empty() {
+                "".to_string()
+            } else {
+                format!(" (revision: {revision})")
+            };
+            let _guard = self.failed_test(&revision, path, cmd, stderr);
+            for error in errors {
+                gha_error(error, &path.display().to_string(), &revision);
+            }
+        }
+        Ok(())
     }
 }
